# swagger2/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  version: '0.1.0'
  title: Boozify Inventory
  description: An API to interact with bars serving beers
host: localhost:8070
basePath: /api
schemes:
  - http
paths:
  /beers:
    post:
      description: create a beer
      operationId: createBeer
      consumes:
        - application/json
      parameters:
        - name: beer
          in: body
          required: true
          schema:
            $ref: '#/definitions/Beer'
      responses:
        '201':
          description: beer created
          schema:
            type: object
    get:
      description: get a list of all beers
      operationId: getBeers
      produces:
        - application/json
      responses:
        '200':
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/Beer'

  /beers/{name}:
    get:
      description: get a beer
      operationId: getBeer
      produces:
        - application/json
      parameters:
        - name: name
          in: path
          required: true
          type: string
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/Beer'
    patch:
      description: update a beer's information
      operationId: updateBeer
      consumes:
        - application/json
      parameters:
        - name: name
          in: path
          required: true
          type: string
        - name: beer
          in: body
          required: true
          schema:
            $ref: '#/definitions/Beer'
      responses:
        '200':
          description: beer updated
    delete:
      description: delete a beer
      operationId: deleteBeer
      parameters:
        - name: name
          in: path
          required: true
          type: string
      responses:
        "204":
          description: beer deleted

  /bars:
    post:
      description: create a bar
      operationId: createBar
      consumes:
        - application/json
      parameters:
        - name: bar
          in: body
          required: true
          schema:
            $ref: '#/definitions/Bar'
      responses:
        '201':
          description: bar created
          schema:
            type: object
    get:
      description: get a list of all bars
      operationId: getBars
      produces:
        - application/json
      responses:
        '200':
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/Bar'

  /bars/{name}:
    get:
      description: get a bar
      operationId: getBar
      produces:
        - application/json
      parameters:
        - name: name
          in: path
          required: true
          type: string
      responses:
        '200':
          description: success
          schema:
            $ref: '#/definitions/Bar'
    patch:
      description: update a bar's information
      operationId: updateBar
      consumes:
        - application/json
      parameters:
        - name: name
          in: path
          required: true
          type: string
        - name: bar
          in: body
          required: true
          schema:
            $ref: '#/definitions/Bar'
      responses:
        '200':
          description:  bar updated
    delete:
      description: delete a bar
      operationId: deleteBar
      parameters:
        - name: name
          in: path
          required: true
          type: string
      responses:
        "204":
          description: bar deleted

  /inventory/beer/{name}:
    get:
      description: get a list of bars serving given beer
      operationId: getBarsServingBeer
      produces:
        - application/json
      parameters:
        - name: name
          in: path
          required: true
          type: string
      responses:
        "200":
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/Bar'

  /inventory/bar/{name}:
    get:
      description: get a list of beers served by given bar
      operationId: getBeersServedByBar
      produces:
        - application/json
      parameters:
        - name: name
          in: path
          required: true
          type: string
      responses:
        "200":
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/Beer'

  /inventory/bar/{name}/{beer}:
    post:
      description: add a given beer to the inventory of a given bar
      operationId: addBeerToBar
      parameters:
        - name: name
          in: path
          required: true
          type: string
        - name: beer
          in: path
          required: true
          type: string
      responses:
        "200":
          description: added beer to bar's inventory

definitions:
  Beer:
    type: object
    properties:
      name:
        type: string
      type:
        type: string
      alcohol:
        type: integer

  Bar:
    type: object
    properties:
      name:
        type: string
      address:
        type: string